Fix tabs and DatasetMembersView
Dataset, Request and Project views each has differet set of tabs, so it
is not a good idea to reuse one tabs.html template for all of them.
Currently I separated tabs.html only for Dataset.

Also refactored DatasetMebersView, wich acts like a list view, bet
extends DatasetDetailsView. Changed it to ListView as it supposed to be.